requirements.txt
flask
openai-whisper
torch
textblob
ffmpeg-python